+++ lib/src/builder/field_builder.dart
@@ -1,5 +1,48 @@
-part of '../models/field_instace_management/form_manager.dart';
+part of 'form_manager.dart';
 
+/// A widget that automatically rebuilds the interface based on changes
+/// to the state of a [FieldController].
+///
+/// The [FieldBuilder] is similar to a `ValueListenableBuilder`, but is
+/// specific for managing Formy fields. It watches a [FieldController]
+/// and rebuilds the widget whenever the field state changes.
+///
+/// This widget is useful to build complex interfaces that depend on the
+/// state of a form field.
+///
+/// ## Properties
+///
+/// * [field]: The [FieldController] this widget will watch.
+/// * [child]: A static widget that will not be rebuilt when the field state changes.
+/// * [builder]: A function that returns the widget to be rebuilt. It receives:
+///   * `context`: The [BuildContext] from the widget tree.
+///   * `field`: The [FieldController] being observed.
+///   * `child`: The static widget defined in the `child` property.
+/// * [buildWhen]: An optional function that returns `true` if the widget
+///   should rebuild. It receives:
+///   * `oldState`: The previous [FieldState].
+///   * `currentState`: The current [FieldState].
+///
+/// ## Example
+/// ```dart
+/// FieldBuilder<String>(
+///   field: FieldController(key: 'key'),
+///   builder: (context, field, child) {
+///     return TextField(
+///       controller: TextEditingController(text: field.value),
+///       decoration: InputDecoration(
+///         labelText: 'Name',
+///         errorText: field.firstError,
+///       ),
+///       onChanged: field.update,
+///     );
+///   },
+/// )
+/// ```
+///
+/// ## See also
+///
+/// * [FieldController], which manages the state of a field.
 class FieldBuilder<T> extends FormyBuilder<FieldController<T>, FieldState<T>> {
   const FieldBuilder({
     super.key,
@@ -9,21 +52,12 @@ class FieldBuilder<T> extends FormyBuilder<FieldController<T>, FieldState<T>> {
     required this.builder,
   });
 
+  ///A function that returns the widget to be rebuilt.
   final Widget Function(
       BuildContext context, FieldController<T> field, Widget? child) builder;
 
   @override
   State<StatefulWidget> createState() => _FieldBuilder<T>();
-
-  @override
-  void insertIntoFormManager() {
-    FormManager.instance.insertField(field);
-  }
-
-  @override
-  void removeFromFormManager() {
-    FormManager.instance._removeField(field);
-  }
 }
 
 class _FieldBuilder<T>